tests/wpt/web-platform-tests/navigation-api/navigate-event/navigate-anchor-same-origin-cross-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<a id="a" href="foo.html"></a>
<script>
async_test(t => {
  navigation.onnavigate = t.step_func_done(e => {
    assert_equals(e.navigationType, "push");
    assert_true(e.cancelable);
    assert_true(e.canTransition);
    assert_false(e.userInitiated);
    assert_false(e.hashChange);
    assert_equals(e.formData, null);
    assert_equals(e.downloadRequest, null);
    assert_equals(new URL(e.destination.url).pathname,
                  "/navigation-api/navigate-event/foo.html");
    assert_false(e.destination.sameDocument);
    assert_equals(e.destination.key, null);
    assert_equals(e.destination.id, null);
    assert_equals(e.destination.index, -1);
    e.preventDefault();
  });
  a.click();
}, "<a> cross-document (but same-origin) navigate event");
</script>
