LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/navigate-anchor-same-origin-cross-document-expected.txt


FAIL <a> cross-document (but same-origin) navigate event assert_equals: expected (object) Element node <a id="a" href="foo.html"></a> but got (undefined) undefined
